Make email login it's own component
